+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+- ⚙ [2017-05-31] `engine/containers/numberGenerator`: update to v4.1.0.
+  - ⚙ Update `hasEnteredToNumericLine` to add jump effect.
 - ⚙ [2017-05-31] `engine/containers/mazeEngineGenerator`: update to v4.1.0.
   - ⚙ Animate the number on `MazePathOverflow` error.
   - ⚙ Update maze position to match new space-in-between the numeric line requirement.